requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
requests~=2.28.2
pymysql~=1.0.2
pikepdf~=7.1.1
pytest~=7.3.1
selenium~=4.9.1
pyotp~=2.8.0